docker-setup/requirements.txt
click==7.1.1
dnspython==1.16.0
kafka-python==2.0.1
mysql-connector-python==8.0.19
protobuf==3.6.1
python-dateutil==2.8.1
python-dotenv==0.12.0
shellescape==3.8.1
six==1.14.0
